# perf(rust, polars):  use iterator instead of loop `polars_io::csv::parser::skip_condition`