@ritchie46 2022-10-10
Nice one. Just shows how often a minor code snippet is executed. :)

> Sidenote: I re-used behavior from the old implementation but there is actually a potential OOB if the predicate does not satisfy any of the remaining bytes in the input buffer. Since input.len() is returned in those cases, the consecutive line will trigger an OOB. (This was also the case in the old implementation.) I am unsure how to fix, since a pointer has to be returned...

Right.. We could return an `Option`. by calling `input.get(read..)`. We already do a bounds check as we do a normal `indexing []` operation, so we might as well use a better OOB path than a panic.